Change int to int8 in signals mode, x4 more efficientgit add .!

+++ headers/signals.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   signals.h                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: egualand <egualand@student.42firenze.it    +#+  +:+       +#+        */
+/*   By: craimond <bomboclat@bidol.juis>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/01 16:04:13 by craimond          #+#    #+#             */
-/*   Updated: 2024/02/10 17:40:32 by egualand         ###   ########.fr       */
+/*   Updated: 2024/02/10 23:28:54 by craimond         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,6 +18,6 @@
 # define S_COMMAND 2
 # define S_SILENT 3
 
-void set_signals(int mode);
+void set_signals(int8_t mode);
 
 #endif
